streaming-compare/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
pyspark==3.4.1
apache-flink==1.18.3
pandas==2.0.3
matplotlib==3.7.1
python-dateutil==2.8.2
